% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPipeSearch.R
\name{autoPipeSearch}
\alias{autoPipeSearch}
\title{autoPipeSearch function}
\usage{
autoPipeSearch(
  x,
  ttime = FALSE,
  blocks = NULL,
  search.term = NULL,
  hits = 2,
  extracturls = FALSE,
  breakloop = 100,
  messages = FALSE
)
}
\arguments{
\item{x}{url, list of urls.}

\item{ttime}{checks if data covers reported turnout over election day (TRUE) or not (FALSE).  ttime is valid when blocks=NULL}

\item{blocks}{vector of blocks to be used to construct a pipe. blocks override ttime parameters. By default, c("listURLextractor()", "rowURLextractor('sayt izbiratel`noy komissii sub`yekta')", "rowURLextractor('Itogi golosovaniya|Rezul`taty vyborov')").}

\item{search.term}{once the stop "term" is detected the algorithm stops building the pipe (by default, UIK|uik|uchastok).}

\item{hits}{number of times the stop word is "hit" by the algorithm. Each time the stop word is "hit" extra block is added to the pipe (by default, hits=2).}

\item{extracturls}{urls are extracted using the pipe or the list of pipes (by default, extracturls=FALSE).}

\item{breakloop}{maximum number of iterations for optimal path search (by default, breakloop=100).}

\item{messages}{display progress messages (messages = FALSE).}
}
\value{
Returns the list elections and pipe.table objects.
}
\description{
This function automatically constructs pipe for a given election.
}
\examples{
library(CECscraper)
library(dplyr)
library(rvest)
 webpage <- read_html(system.file("elections_primorye.html", package="CECscraper"))

 #Example 1
 wterrit <- webpage \%>\% fileURLextractor(hashid = FALSE)
 #uik_url1 <- wterrit[1,]\%>\%autoPipeSearch(extracturls=TRUE, ttime=FALSE)
 uik_url1 <- wterrit[1,]\%>\%autoPipeSearch(ttime=TRUE)
 uik_url2 <- uik_url1\%>\%execPipeList()
 #uik_url3 <- uik_url2\%>\%dataBuilder(ttime=FALSE)
  uik_url3 <- uik_url2\%>\%dataBuilder(ttime=TRUE)

 #Example 2
 uik_url1 <- wterrit[1,]\%>\%autoPipeSearch(blocks=c("listURLextractor()",
                                                  "rowURLextractor('sayt izbiratel`noy komissii sub`yekta')",
                                                  "rowURLextractor('Svodnaya tablitsa')"),  hits=3, search.term="Svodnaya tablitsa")
 uik_url2 <- uik_url1\%>\%
            execPipeList()\%>\%
            dataBuilder(typedata = "fast")

 #Example 3
 uik_url1 <- wterrit[1,]\%>\%autoPipeSearch(blocks=c("listURLextractor()",
                                                  "rowURLextractor('sayt izbiratel`noy komissii sub`yekta')",
                                                  "rowURLextractor('Dannyye ob otkrytii pomeshcheniy')"),  hits=3, search.term="Dannyye ob otkrytii pomeshcheniy")
 uik_url2 <- uik_url1\%>\%
            execPipeList()\%>\%
            dataBuilder(ttime = TRUE, typedata = "fast")

 #Example 4
 uik_url1 <- wterrit[1,]\%>\%autoPipeSearch(blocks=c("listURLextractor()",
                                               "rowURLextractor('sayt izbiratel`noy komissii sub`yekta')",
                                               "rowURLextractor('Svedeniya o kandidatakh')"),  hits=1, search.term="Svedeniya o kandidatakh")
 uik_url2 <- uik_url1\%>\%
             execPipeList()\%>\%
             scrapeCandidates(tabextract = NULL, savetodir = "")
}
