% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMerger.R
\name{DataMerger}
\alias{DataMerger}
\title{DataMerger function}
\usage{
DataMerger(x, byrow = TRUE)
}
\arguments{
\item{x}{the list of data objects.}

\item{byrow}{the list is merged by row or by column.  If byrow==FALSE only two objects must be put into the list.}
}
\value{
Returns the data.frame object with urls and links.
}
\description{
This function merges the list of data objects together.
}
\examples{
library(CECscraper)

#Example 1
murl<-"https://tinyurl.com/yy6roo3g"
uiks<-MenuLinkExtractor(MenuLinkExtractor(MenuLinkExtractor(murl))[1:5,])

uiks_turnout<-PageLinkExtractor(uiks, "Dannyye ob otkrytii pomeshcheniy dlya golosovaniya")
uiks_voting<-PageLinkExtractor(uiks, "Rezul'taty vyborov|vyborov po odnomandatnomu \\\\(mnogomandatnomu\\\\) okrugu")

uiks_turnout_data<-DataBuilder(uiks_turnout, bylevel="level1", ttime=TRUE)
uiks_voting_data<-DataBuilder(uiks_voting, bylevel="level1", ttime=FALSE)
uiks_merged<-DataMerger(list(uiks_voting_data,uiks_turnout_data), byrow=FALSE)

#Example 2
# library(dplyr)
# murl="https://tinyurl.com/y369jngp"
#fast_downloadT<-PageLinkExtractor(murl, "Predvaritel'nyye svedeniya ob uchastii izbirateley")\%>\%
#                 MenuLinkExtractor()\%>\%MenuLinkExtractor()\%>\%sample_n(100, replace = FALSE)\%>\%
#                 PageLinkExtractor("sayt izbiratel'noy komissii sub\\"yekta Rossiyskoy Federatsii")\%>\%
#                 DataBuilder(typedata="fast", bylevel="level2", ttime=TRUE)\%>\%DataMerger()
}
