% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFTformatter.R
\name{EFTformatter}
\alias{EFTformatter}
\title{EFTformatter function}
\usage{
EFTformatter(x, Nvalid = "CEC", levels = TRUE)
}
\arguments{
\item{x}{the object from Databuilder/DataMerger function.}

\item{Nvalid}{valid votes computed using either "CEC" (using the Central Election Comission's formula); "AllVotes" (total number of votes cast for all candidates).}

\item{levels}{add information on levels (TRUE) or not (FALSE) to the output.}
}
\value{
Returns the data.frame object with three variables: "Votes" (number of ballots for the leading parties or candidates), "Nvalid" (number of casted ballots), "Voters" (number of eligible voters).
}
\description{
This function reformats the data for its use with Election Forensics Toolkit.
}
\examples{
library(CECscraper)

Rurls<-"https://tinyurl.com/yy6roo3g"
uiks<-MenuLinkExtractor(MenuLinkExtractor(MenuLinkExtractor(Rurls))[1:5,])
uiks_voting<-PageLinkExtractor(uiks, "Rezul'taty vyborov|vyborov po odnomandatnomu \\\\(mnogomandatnomu\\\\) okrugu")
uiks_voting_data<-DataBuilder(uiks_voting, bylevel="level1", ttime=FALSE)
uiksEFT<-EFTformatter(uiks_voting_data)
}
