% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execPipeList.R
\name{execPipeList}
\alias{execPipeList}
\title{execPipeList function}
\usage{
execPipeList(x, messages = TRUE)
}
\arguments{
\item{x}{autoPipeSearch object containing pipe.table that can be manually modified if desired.}

\item{messages}{display progress messages (messages = FALSE).}
}
\value{
Extracts the data via a pipe or a list of pipes and saves it inside the elections object.
}
\description{
This function extracts URLs from the pipe saved in the autoPipeSearch object.
}
\examples{
library(CECscraper)
library(dplyr)
library(rvest)

webpage <- read_html(system.file("elections_primorye.html", package="CECscraper"))

 wterrit <- webpage \%>\% fileURLextractor(hashid = FALSE)
 uik_url1 <- wterrit[1,]\%>\%autoPipeSearch(extracturls=TRUE, ttime=TRUE)
 uik_url2 <- uik_url1\%>\%execPipeList()
 uik_url3 <- uik_url2\%>\%dataBuilder(ttime=TRUE)
}
