% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listURLextractor.R
\name{listURLextractor}
\alias{listURLextractor}
\title{listURLextractor function}
\usage{
listURLextractor(x)
}
\arguments{
\item{x}{link, list of links or data frame with links}
}
\value{
Returns the dataframe object with urls, links.
}
\description{
This function extracts links and urls from the lists/menus.
}
\examples{
#Example
library(CECscraper)

murl <- "https://tinyurl.com/yy6roo3g"
oiks <- listURLextractor(murl)
rayons <- listURLextractor(listURLextractor(murl))
uiks <- listURLextractor(listURLextractor(listURLextractor(murl))[1:5,])
}
