% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MenuLinkExtractor.R
\name{MenuLinkExtractor}
\alias{MenuLinkExtractor}
\title{MenuLinkExtractor function}
\usage{
MenuLinkExtractor(x)
}
\arguments{
\item{x}{link, list of links or data frame with links}
}
\value{
Returns the dataframe object with urls, links.
}
\description{
This function extracts the links from the dropdown menu.
}
\examples{
#Example
library(CECscraper)

murl<-"https://tinyurl.com/yy6roo3g"
oiks<-MenuLinkExtractor(murl)
rayons<-MenuLinkExtractor(MenuLinkExtractor(murl))
uiks<-MenuLinkExtractor(MenuLinkExtractor(MenuLinkExtractor(murl))[1:5,])
}
