% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageLinkExtractor.R
\name{PageLinkExtractor}
\alias{PageLinkExtractor}
\title{PageLinkExtractor function}
\usage{
PageLinkExtractor(x, item, select = 1)
}
\arguments{
\item{x}{the list of urls or MenuListExtractor object.}

\item{item}{link/expression that contains the url of interest.}

\item{select}{select id of a single match among several identified by the algorithm.}
}
\value{
Returns the data.frame object containing levels, links and urls.
}
\description{
This function extracts the links and urls from the page.
}
\examples{
library(CECscraper)

murl<-"https://tinyurl.com/yy6roo3g"
oiks<-MenuLinkExtractor(murl)
rayons<-MenuLinkExtractor(MenuLinkExtractor(murl))
#rayons<-MenuLinkExtractor(murl)\%>\%MenuLinkExtractor()
uiks<-MenuLinkExtractor(MenuLinkExtractor(MenuLinkExtractor(murl))[1:5,])

uiks_turnout<-PageLinkExtractor(uiks, "Dannyye ob otkrytii pomeshcheniy dlya golosovaniya")
uiks_voting<-PageLinkExtractor(uiks, "Rezul'taty vyborov|vyborov po odnomandatnomu \\\\(mnogomandatnomu\\\\) okrugu")
}
