% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowURLextractor.R
\name{rowURLextractor}
\alias{rowURLextractor}
\title{rowURLextractor function}
\usage{
rowURLextractor(x, item, select = 1)
}
\arguments{
\item{x}{the list of urls or MenuListExtractor object.}

\item{item}{link/expression that contains the url of interest.}

\item{select}{select id of a single match among several identified by the algorithm.}
}
\value{
Returns the data.frame object containing levels, links and urls.
}
\description{
This function extracts the links and urls from the page.
}
\examples{
library(CECscraper)

murl<-"https://tinyurl.com/yy6roo3g"
oiks<-listURLextractor(murl)
rayons<-listURLextractor(listURLextractor(murl))
#rayons<-listURLextractor(murl) \%>\% listURLextractor()
uiks<-listURLextractor(listURLextractor(listURLextractor(murl))[1:5,])

uiks_turnout<-rowURLextractor(uiks, "Dannyye ob otkrytii pomeshcheniy dlya golosovaniya")
uiks_voting<-rowURLextractor(uiks, "Rezul'taty vyborov|vyborov po odnomandatnomu \\\\(mnogomandatnomu\\\\) okrugu")
}
