% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transliterate.R
\name{transliterate}
\alias{transliterate}
\title{Transliteration function.}
\usage{
transliterate(v)
}
\arguments{
\item{v}{expression}
}
\value{
Returns transliterated expression.
}
\description{
This function transliterates expression from Cyrillic to Latin.
}
\examples{
library(CECscraper)

transliterate("enter cyrillic text here")
}
