% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBuilder.R
\name{dataBuilder}
\alias{dataBuilder}
\title{dataBuilder function}
\usage{
dataBuilder(x, bylevel = NULL, ttime = FALSE, typedata = "slow",
  dnames = FALSE, tabextract = NULL, savetodir = "",
  messages = TRUE)
}
\arguments{
\item{x}{link/list/data frame with links.}

\item{bylevel}{define data subsets for an output (i.e. single-member districts); if bylevel=NULL the whole data set is returned.}

\item{ttime}{checks if extracted data covers reported turnout over election day (TRUE) or not (FALSE).}

\item{typedata}{"slow" if the data is extracted from "rezultaty vyborov" link (slow approach); "fast" if the data is extracted from "svodnya tablitsa" (pivot table) (fast approach).}

\item{dnames}{assign original labels to column names(TRUE).}

\item{tabextract}{select the table number to extract in order to override the table selection algorithm.}

\item{savetodir}{save html data files to specified directory, i.e. "C:/Documents".}

\item{messages}{display progress messages (TRUE).}
}
\value{
The list containing the following parameters:
\itemize{
  \item data - data frame or the list of data frames
  \item ttime - input parameter
  \item dnames - input parameter
  \item bylevel - input parameter
  \item retreivaldate - time, date of data retrieval
}
}
\description{
This function builds the data frame from the webpage or listURLextractor/rowURLextractor objects.
}
\examples{
library(CECscraper)

#Example 1
murl<-"https://tinyurl.com/yy6roo3g"
uiks<-listURLextractor(listURLextractor(listURLextractor(murl))[1:5,])

uiks_turnout<-rowURLextractor(uiks, "Dannyye ob otkrytii pomeshcheniy dlya golosovaniya")
uiks_voting<-rowURLextractor(uiks, "Rezul`taty vyborov|vyborov po odnomandatnomu \\\\(mnogomandatnomu\\\\) okrugu")

uiks_turnout_data<-dataBuilder(uiks_turnout, bylevel="level1", ttime=TRUE)
uiks_voting_data<-dataBuilder(uiks_voting, bylevel="level1", ttime=FALSE)
uiks_merged<-dataMerger(list(uiks_voting_data,uiks_turnout_data), byrow=FALSE)

#Example 2
# library(dplyr)
# murl="https://tinyurl.com/y369jngp"
#fast_downloadT<-rowURLextractor(murl, "Predvaritel'nyye svedeniya ob uchastii izbirateley")\%>\%
#                 listURLextractor()\%>\%listURLextractor()\%>\%sample_n(100, replace = FALSE)\%>\%
#                 rowURLextractor("sayt izbiratel'noy komissii sub\\"yekta Rossiyskoy Federatsii")\%>\%
#                 dataBuilder(typedata="fast", bylevel="level2", ttime=TRUE)\%>\%dataMerger()
}
