% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMerger.R
\name{dataMerger}
\alias{dataMerger}
\title{dataMerger function}
\usage{
dataMerger(x, byrow = TRUE)
}
\arguments{
\item{x}{list of data objects.}

\item{byrow}{list is merged by row or by column.  If byrow==FALSE only two objects are merged into a list.}
}
\value{
Returns the data.frame object with urls and links.
}
\description{
This function merges the list of data objects together.
}
\examples{
library(CECscraper)

#Example 1
murl<-"http://notelections.online/region/izbirkom?action=show&vrn=27720002327736&region=77&prver=0&pronetvd=null"
uiks <- listURLextractor(
            rowURLextractor(
               listURLextractor(
                   listURLextractor(murl))[1:5,], 'sayt izbiratel`noy komissii sub`yekta'))

uiks_turnout<-rowURLextractor(uiks, "Dannyye ob otkrytii pomeshcheniy dlya golosovaniya")
uiks_voting<-rowURLextractor(uiks, "Rezul`taty vyborov|vyborov po odnomandatnomu \\\\(mnogomandatnomu\\\\) okrugu")

uiks_turnout_data<-dataBuilder(uiks_turnout, bylevel="level1", ttime=TRUE)
uiks_voting_data<-dataBuilder(uiks_voting, bylevel="level1", ttime=FALSE)
uiks_merged<-dataMerger(list(uiks_voting_data,uiks_turnout_data), byrow=FALSE)

#Example 2
# library(dplyr)
# murl="http://notelections.online/region/izbirkom?action=show&global=1&vrn=100100084849062&region=0&prver=0&pronetvd=null"
#fast_downloadT<-rowURLextractor(murl, "Predvaritel`nyye svedeniya ob uchastii izbirateley")\%>\%
#                 listURLextractor()\%>\%listURLextractor()\%>\%sample_n(100, replace = FALSE)\%>\%
#                 rowURLextractor("sayt izbiratel`noy komissii sub`yekta Rossiyskoy Federatsii")\%>\%
#                 dataBuilder(typedata="fast", bylevel="level2", ttime=TRUE)\%>\%dataMerger()
}
