% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eftFormatter.R
\name{eftFormatter}
\alias{eftFormatter}
\title{eftFormatter function}
\usage{
eftFormatter(x, Nvalid = "CEC", levels = TRUE)
}
\arguments{
\item{x}{the object from Databuilder/DataMerger function.}

\item{Nvalid}{valid votes computed using either "CEC" (using the Central Election Comission's formula); "AllVotes" (total number of votes cast for all candidates).}

\item{levels}{add information on levels (TRUE) or not (FALSE) to the output.}
}
\value{
Returns the data.frame object with three variables: "Votes" (number of ballots for the leading parties or candidates), "Nvalid" (number of casted ballots), "Voters" (number of eligible voters).
}
\description{
This function reformats the data for its use with Election Forensics Toolkit.
}
\examples{
library(CECscraper)

 murl <- "http://notelections.online/region/izbirkom?action=show&vrn=27720002327736&region=77&prver=0&pronetvd=null"
uiks <- listURLextractor(
            rowURLextractor(
               listURLextractor(
                   listURLextractor(murl))[1:5,], 'sayt izbiratel`noy komissii sub`yekta'))
uiks_voting <- rowURLextractor(uiks, "Rezul`taty vyborov|vyborov po odnomandatnomu \\\\(mnogomandatnomu\\\\) okrugu")
uiks_voting_data <- dataBuilder(uiks_voting, bylevel="level1", ttime=FALSE)
uiksEFT <- eftFormatter(uiks_voting_data)
}
