% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileURLextractor.R
\name{fileURLextractor}
\alias{fileURLextractor}
\title{fileURLextractor function}
\usage{
fileURLextractor(html_file, tabextract = NULL, hashid = FALSE)
}
\arguments{
\item{html_file}{html file object.}

\item{tabextract}{select the table number to extract in order to override the table selection algorithm.}

\item{hashid}{generate md5 hash id for an election.}
}
\value{
Returns the data.frame object containing levels, links and urls of elections.
}
\description{
This function extracts the links and urls from the html page of the Central Election Commission.
}
\examples{
library(CECscraper)
library(rvest)

webpage <- read_html(system.file("elections_primorye.html", package="CECscraper"))

wterrit <- webpage \%>\% fileURLextractor(hashid = TRUE)
wterrit$webscrape[1] <- FALSE
res_territ <- wterrit \%>\% rowURLextractor("Rezul'taty vyborov") \%>\%
listURLextractor() \%>\%
dataBuilder(ttime = FALSE, typedata = "slow", bylevel = "level3")

}
