% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listURLextractor.R
\name{listURLextractor}
\alias{listURLextractor}
\title{listURLextractor function}
\usage{
listURLextractor(x, messages = TRUE)
}
\arguments{
\item{x}{link, list of links or data frame with links}

\item{messages}{display progress messages (TRUE).}
}
\value{
Returns the dataframe object with urls, links.
}
\description{
This function extracts links and urls from the lists/menus.
}
\examples{
#Example
library(CECscraper)

murl <- "http://notelections.online/region/izbirkom?action=show&vrn=27720002327736&region=77&prver=0&pronetvd=null"
oiks <- listURLextractor(murl)
rayons <- listURLextractor(listURLextractor(murl))
uiks <- listURLextractor(
            rowURLextractor(
               listURLextractor(
                   listURLextractor(murl))[1:5,], 'sayt izbiratel`noy komissii sub`yekta'))
}
