% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowURLextractor.R
\name{rowURLextractor}
\alias{rowURLextractor}
\title{rowURLextractor function}
\usage{
rowURLextractor(x, item, select = 1, messages = TRUE)
}
\arguments{
\item{x}{list of urls or MenuListExtractor object.}

\item{item}{link/expression that contains the url of interest.}

\item{select}{if more than one matched rows are found, define which one to work with.}

\item{messages}{display progress messages (TRUE).}
}
\value{
Returns the data.frame object containing levels, links and urls.
}
\description{
This function extracts link and url from the webpage's row defined by specific expression.
}
