% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeCandidates.R
\name{scrapeCandidates}
\alias{scrapeCandidates}
\title{scrapeCandidates function}
\usage{
scrapeCandidates(x, tabextract = NULL, savetodir = "")
}
\arguments{
\item{x}{url, list of urls.}

\item{tabextract}{select the table number to extract in order to override the table selection algorithm.}

\item{savetodir}{save downloaded html files to directory, i.e. "C:/Documents".}
}
\value{
Returns the list containing the data.frame object and retrieval date.
}
\description{
This function extracts the candidate-related data from the webpages.
}
\examples{
library(CECscraper)

url.moscow.candidates="https://tinyurl.com/yyhtfnwg"
moscow.candidates<-scrapeCandidates(url.moscow.candidates)
}
