% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeCandidates.R
\name{scrapeCandidates}
\alias{scrapeCandidates}
\title{scrapeCandidates function}
\usage{
scrapeCandidates(x, tabextract = NULL, savetodir = "", messages = TRUE)
}
\arguments{
\item{x}{url, list of urls.}

\item{tabextract}{select the table number to extract in order to override the table selection algorithm.}

\item{savetodir}{save downloaded html files to directory, i.e. "C:/Documents".}

\item{messages}{display progress messages (TRUE).}
}
\value{
Returns the list containing the data.frame object and retrieval date.
}
\description{
This function extracts the candidate-related data from the webpages.
}
\examples{
library(CECscraper)

url.moscow.candidates="http://notelections.online/region/region/moscow_city?action=show&root=1&tvd=27720002327740&vrn=27720002327736&region=77&global=null&sub_region=0&prver=0&pronetvd=null&vibid=27720002327736&type=220"
moscow.candidates<-scrapeCandidates(url.moscow.candidates)
}
