% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapeCandidates.R
\name{ScrapeCandidates}
\alias{ScrapeCandidates}
\title{ScrapeCandidates function}
\usage{
ScrapeCandidates(x, savetodir = "")
}
\arguments{
\item{x}{url, list of urls}

\item{savetodir}{save html downloaded data files in directory, i.e. "C:/Documents"}
}
\value{
Returns the list with data.frame object and retrieval date.
}
\description{
This function extracts the candidate-related data information from the webpages.
}
\examples{
library(CECscraper)

url.moscow.candidates="https://tinyurl.com/yyhtfnwg"
moscow.candidates<-ScrapeCandidates(url.moscow.candidates)
}
