% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transliterate.R
\name{Transliterate}
\alias{Transliterate}
\title{Transliteration function.}
\usage{
Transliterate(v)
}
\arguments{
\item{v}{expression}
}
\value{
Returns transliterated expression.
}
\description{
This function transliterates expression from Cyrillic to Latin.
}
\examples{
library(CECscraper)

Transliterate("enter cyrillic text here")
}
